fix: modify exporter to be 8.8 compatible
