+++ parent/pom.xml
@@ -69,6 +69,7 @@
     <version.surefire>3.5.6</version.surefire>
     <version.jackson>2.21.4</version.jackson>
     <version.junit>5.13.4</version.junit>
+    <version.junit-pioneer>2.3.0</version.junit-pioneer>
     <version.junit4>4.13.2</version.junit4>
     <version.log4j>2.25.4</version.log4j>
     <version.minlog>1.3.1</version.minlog>
@@ -1008,6 +1009,12 @@
         <version>${version.junit}</version>
         <type>pom</type>
         <scope>import</scope>
+      </dependency>
+
+      <dependency>
+        <groupId>org.junit-pioneer</groupId>
+        <artifactId>junit-pioneer</artifactId>
+        <version>${version.junit-pioneer}</version>
       </dependency>
 
       <dependency>
